analyticaldashboard/backend/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
numpy>=1.21.0
python-dotenv>=1.0.0
requests>=2.28.0
gunicorn>=20.1.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
google-generativeai>=0.3.0
nltk>=3.8